requirements.txt
# Core Web Framework
streamlit==1.28.1

# AI and Machine Learning
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
huggingface-hub==0.17.3

# Vector Database
chromadb==0.4.15

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6

# Data Processing
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0

# Web Requests and APIs
requests==2.31.0
urllib3==2.0.7

# Optional: FastAPI (for future backend)
fastapi==0.104.1
uvicorn==0.24.0

# Utilities
python-dotenv==1.0.0
pathlib==1.0.1
typing-extensions==4.8.0

# Development Tools (Optional)
pytest==7.4.2
black==23.9.1
flake8==6.1.0